int preco

+++ src/main/java/com/example/SportsJerseyKingdom/Models/Camisa.java
@@ -2,8 +2,8 @@ package com.example.SportsJerseyKingdom.Models;
 
 import java.util.Random;
 
-public record Camisa(Long id, String nome, String marca, String tamanho, float preco,String foto) {
-    public Camisa(Long id, String nome, String marca, String tamanho, float preco, String foto){
+public record Camisa(Long id, String nome, String marca, String tamanho, int preco,String foto) {
+    public Camisa(Long id, String nome, String marca, String tamanho, int preco, String foto){
         //o abs pega só o numero absoluto ignorando o sinal
         var key = (id!= null) ? id : Math.abs(new Random().nextLong() );// se já tem id, n é preciso passar um aleatorio pois pega o que já tem
         this.id = key; //esses comportamento do id é o que difere do record
